Refactoring the various infinite baffle/half-space simulations of radially symmetric horns.
